requirements.txt
Pillow >= 3
